requirements.txt
streamlit>=1.36.0
google-generativeai>=0.8.2
python-dotenv>=1.0.1
pdf2image>=1.17.0
pillow>=10.3.0
json-repair>=0.10.0
PyPDF2>=3.0.1
pathlib
argparse
openpyxl